clean up

+++ remove-bg.ps1
@@ -1,21 +1,31 @@
 param([int]$batches,
 [int]$concurrency,
 [string]$target,
-[int]$fps)
+[int]$fps,
+[bool]$skip_create_frames)
 
-ffmpeg -i $target -vf fps=$fps %d.jpg
+mkdir ./tmp -Force
+cp $target ./tmp
+cd tmp
 
 $folder = Get-Location
+
+# turn video into frames
+if(-not $skip_create_frames){
+    ffmpeg -i $target -vf fps=$fps %d.jpg
+}
 
 # delete any failed png conversions
 ls *.png | where Length -eq 0 | del
 
 $d = 0;
 
+# process all the frames into batches
 Get-ChildItem | 
+    Where-Object Name -imatch "jpg" |
     ForEach-Object { 
         New-Object PSObject  -property @{
-            Test= -Not (Test-Path -Path ($_.name -replace "jpg", ".out.png") ); 
+            Test= -not (Test-Path -Path ($_.name -replace "jpg", ".out.png") ); 
             Path=$_.Name;
             Number=$d++
         } } | 
@@ -40,24 +50,42 @@ Get-ChildItem |
 
 get-job | Wait-Job
 
-# do a second pass to pick up the failed ones
-$failed = ls | where Length -eq 0 |  %{ $_.Name -replace ".out.png", "" }
 
-mkdir failed -Force
-$failed | ForEach-Object{ Copy-Item "$_.jpg" failed }
+# retry failed ones
+1..2 | % { 
 
-C:/Windows/System32/cmd.exe /K "$env:anaconda\\Scripts\\activate.bat $env:anaconda && cd $env:rembg && python -m src.rembg.cmd.cli -p $folder\failed 2>&1 && exit"
+    # do a second pass to pick up the failed ones
+    $failed = ls *.png | where Length -eq 0 |  %{ $_.Name -replace ".out.png", "" }
 
-Copy-Item $folder/failed/*.png $folder
-Remove-Item $folder/failed/ -Force -Recurse
+    mkdir failed -Force
+    $failed | ForEach-Object{ Copy-Item "$_.jpg" failed }
 
-# if no failed pngs and >100 non failed pngs
-if( (ls *.png | where Length -eq 0).Count -eq 0 `
-        -and (ls *.png | where Length -gt 0).Count -gt 100)
-{
-    del *.jpg
+    if( ($failed | Measure-Object ).count -gt 0 )
+    {
+        C:/Windows/System32/cmd.exe /K "$env:anaconda\\Scripts\\activate.bat $env:anaconda && cd $env:rembg && python -m src.rembg.cmd.cli -p $folder\failed 2>&1 && exit"
+
+        Copy-Item $folder/failed/*.png $folder
+        Remove-Item $folder/failed/ -Force -Recurse
+    }
+ }
+
+#check for every JPG file there is a corresponding non-empty PNG present
+if( (Get-ChildItem | 
+    Where-Object Name -imatch "jpg" |
+    ForEach-Object { 
+        New-Object PSObject  -property @{
+            PngNotCreated = -not (Test-Path -Path ($_.name -replace "jpg", ".out.png") ); 
+            PngEmpty = -not (Test-Path -Path ($_.name -replace "jpg", ".out.png") ) -or (get-item ($_.name -replace "jpg", ".out.png")).Length -eq 0;
+        } } | 
+        Where-Object {$_.PngNotCreated -or $_.PngEmpty }|
+        Measure-Object).Length -eq 0 ) {
+
+            del *.jpg
     
-    ffmpeg -i %d.out.png -vcodec png ($target -replace ".mp4", ".mov")
+            ffmpeg -i %d.out.png -vcodec png ($target -replace ".mp4", ".mov")
+        
+            del *.png
+}
 
-    del *.png
-} 
+# go back
+cd ../
